Initial commit.
Download copies of the 25.01.2016 menus for the Uhlhornsweg campus so that we can test the script without accessing the web site repeatedly.  They're in the "test menu pages" directory.

Put links to the menus' URLs in urls.py for future reference.

Do a quick test run of BeautifulSoup4 in Main.py.