Allow the user to save their email address.
In index.html, add an email entry form that POSTs the user's input
to /email_entry.  Also, display the user's saved email address, if present.

In oauth_example_app.py, route /email_entry to the function add_user_email(),
which updates the user's email address in the database and redirects them
back to the home page.