Add db constraint to prevent saving duplicate menu entries
Enforce a unique combination of date_valid, description, mensa, and
category for each item in the menu_entries table.

In case the scraper is run more than once on the same week's menu, it
should not be able to save the same entry (i.e. identical description,
mensa, date, and category) twice. This constraint will prevent that
from happening (causing the scraper to throw an exception if it tries
to do so).

Of course, if the way the scraper itself processes the data changes,
then it may be possible for the same entry to be saved twice
with different whitespace or something.