Proof of concept: Scrape Uhlhornsweg Aufgabe A using Scrapy
Parse and save the menu for Uhlhornsweg Aufgabe A into our database.
Test: 'cd /vagrant
	   scrapy crawl mensa_spider'
To reset the table , run 'scrapy shell' and execute the expression
'MenuEntry.query.delete()'.  It will return a number indicating how
many rows were deleted.
You can also run other queries to check out the data that was scraped.

As of this commit, the scraper will throw an error if run twice on the
same menu without cleaning the database in between, because the db
schema prohibits saving duplicate items.  This is only a safeguard;
The check for duplicates should also be performed in the
scraper itself.

Create a Scrapy project module in /menuscraper
Define a Scrapy Item to represent menu entries: MenuEntry (items.py)
Define a Pipeline to save MenuEntry Items into the db (pipelines.py)
Define a Spider to download a menu from the website (mensa_spider.py)

The spider has a parse method which calls BeautifulSoup4.  All parsing
logic is contained in a separate file (soup_parser.py)

Configure scrapy settings: Limit concurrent requests to one at a time.
Add our pipeline to ITEM_PIPELINES (settings.py)

Create scrapy.cfg, which tells Scrapy to look inside of the
menuscraper module when it's called from the command line.