Display mensa history on web page at /mensahistory
Quick proof of concept reading scraped data out of the database and
displaying it on a webpage with Flask.

Add mensahistory.html - A basic template with a searchable table
displaying all MenuEntry items stored in the database, sorted by date.

Add searchable-table.js - A little bit of Javascript I found which
enables the user to quickly filter the table by typing into a field.

views.py
Add a route for /mensahistory
Runs a query on MenuEntry and passes the result into the template