Create a short script to run Scrapy
main.py: This short script runs scrapy with the command "scrapy crawl
mensa_spider". This would be suitable to schedule as a cron job, e.g.
once a night after midnight. (You could wait as long as a week in
between scrapes, but a few requests once a day is probably not an
unreasonable amount of traffic to send to the Studentenwerk.)