Improve menu parsing. Add support for tables within tables.
soup_parser.py: Add recursive=False parameter when calling find_all() to
split a Mensa menu table into rows ('tr' tags) and columns ('td' tags).
This avoids problems caused by certain menus, as in the infamous case of
the "Wiener Würstchen" on the menu for Uhlhornsweg Aufgabe A from
15.08.2016. In that case, there was a little baby table embedded inside
of one of the cells of the main menu table.

test menu pages/uhlhornsweg-ausgabe-a-20160815 Weird menu with extra
row. wiener wurstchen.html: This is the HTML page in question which
raised the aforementioned issue.