Add a Scrapy Pipeline to filter duplicate menu entries.
pipelines.py: Introduce a new Pipeline, called FilterDuplicates, which
drops menu entries if they are found to already exist in the db. Rename
the Pipeline formerly known as "DatabasePipeline." It is now called
"SaveToDatabase."

settings.py: FilterDuplicates is configured as the first Pipeline.
SaveToDatabase is configured as the second Pipeline.