Restyle home page, rewrite main text, add my email address
Use DataTables plug-in for jQuery to make the table searchable and
sortable by columns. This replaces the previous solution for sorting the
table, searchable-table.js.

Rewrite main text. Add disclaimer. Justify text to the left. Add a more
prominent link to the official Mensa menu.

Add my email address to the nav bar.