Create second OAuth callback route for rerequesting permissions
This commit is a little bit involved, so bear with me. The main point of
this commit is to allow us to have two distinct OAuth callback routes,
one of which is used when a user logs in, and the other of which is used
when we rerequest permissions we need from a user.

As a secondary matter, we are no longer using the term "reauthorizing",
but rather "rerequesting permissions." This is a bit of a mouthful, but
it's more descriptive.

views.py:
Rename oauth_callback() to oauth_callback_authorize().
Rename oauth_reauthorize() to oauth_rerequest_permissions().

Create a new route, oauth_callback_rerequest_permissions(). THIS is, in
a way, the whole point of this commit. We now have a distinct callback
which is used when rerequesting permissions from a user.

oauth.py:
Rename reauthorize() to rerequest_permissions().
Rename callback() to callback_authorize().
Create callback_rerequest_permissions().

Split get_callback_url() into two functions:
get_callback_url_for_authorize() and
get_callback_url_for_rerequest_permissions().

Refactor callback() to take a redirect_uri (which is sent to Facebook)
as an argument. Create two functions which delegate to it:
callback_authorize() and callback_rerequest_permissions().