"Log in with Facebook to subscribe to alerts for this search"
When a user clicks "Log in with Facebook to receive alerts", pass their
search terms through FB as a parameter in our redirect URL.  When the user
is done logging in, read the search terms out of the URL and save the search.
This way, they don't have to type the search terms in a second time.