Fix Vagrant dev environment.  Upgrade cryptography module
Upgrade from Ubuntu 14 to 16
Upgrade from cryptography 1.2.2 to 1.4
Run apt-get upgrade in bootstrap.sh
Upgrade pip in bootstrap.sh

Backstory:
I tried spinning up a new VM, and it wasn't able to install all of the
python packages that I asked for in requirements.txt, so nothing
worked.  With this commit, it should be working again.